- took changes from banuba_with_agora_master

+++ android/src/main/java/io/agora/rtc/base/MediaObserver.kt
@@ -31,10 +31,8 @@ class MediaObserver(
   }
 
   override fun onMetadataReceived(buffer: ByteArray, uid: Int, timeStampMs: Long) {
-    emit(
-      hashMapOf(
-        "data" to arrayListOf(String(buffer), uid.toUInt().toLong(), timeStampMs)
-      )
-    )
+    emit(hashMapOf(
+      "data" to arrayListOf(String(buffer), uid, timeStampMs)
+    ))
   }
 }
